Move msg service router into handlers package
